producer/requirements.txt
websocket-client==1.8.0
kafka-python==2.0.2
python-dotenv==1.0.1
